+++ admin/src/App.tsx
@@ -4,8 +4,8 @@ import simpleRestProvider from "ra-data-simple-rest";
 import React from "react";
 import { Admin, Resource } from "react-admin";
 
-import authProvider from "./auth/authProvider";
-import httpClient from "./auth/httpClient";
+import authProvider from "./auth/auth-provider";
+import httpClient from "./auth/http-client";
 import UserList from "./components/UserList";
 
 const dataProvider = simpleRestProvider(process.env.REACT_APP_API_URL || "", httpClient);
